+++ app/models/dorsale/billing_machine/invoice.rb
@@ -71,7 +71,7 @@ class Dorsale::BillingMachine::Invoice < ::Dorsale::ApplicationRecord
 
     lines.each do |line|
       line_total = line.total - (line.total * discount_rate)
-      self.vat_amount += (line_total * line.vat_rate / 100.0)
+      self.vat_amount += (line_total * line.vat_rate / 100.0).round(2)
     end
 
     self.total_including_taxes = total_excluding_taxes + vat_amount
@@ -91,7 +91,7 @@ class Dorsale::BillingMachine::Invoice < ::Dorsale::ApplicationRecord
       raise "Invoice has multiple vat rates"
     end
 
-    vat_rates.first || ::Dorsale::BillingMachine::DEFAULT_VAT_RATE
+    vat_rates.first || ::Dorsale::BillingMachine.default_vat_rate
   end
 
   attr_writer :vat_rate