+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Next version
 
+- Billing Machine : fix amounts round
+- Billing Machine : change ::DEFAULT_VAT_RATE to ::default_vat_rate
+- Billing Machine : refactor/improve JS
+
 ## 3.9.8
 
 - Tasks : add increment progress buttons